poprawki do open beta, niby 106 plikow a i tam w pizdu do zrobienia jeszcze ;D

+++ src/main/java/rpg/rpgcore/commands/admin/GiveAkcesoriaCommand.java
@@ -26,17 +26,17 @@ public class GiveAkcesoriaCommand extends CommandAPI {
             return;
         }
 
-        double val1;
-        double val2;
-        double val3;
-        double val4;
+        int val1;
+        int val2;
+        int val3;
+        int val4;
         int val5;
         int lvl;
 
         try {
-            val1 = Double.parseDouble(args[1]);
-            val2 = Double.parseDouble(args[2]);
-            val3 = Double.parseDouble(args[3]);
+            val1 = Integer.parseInt(args[1]);
+            val2 = Integer.parseInt(args[2]);
+            val3 = Integer.parseInt(args[3]);
             lvl = Integer.parseInt(args[6]);
         } catch (NumberFormatException e) {
             player.sendMessage(Utils.format(Utils.SERVERNAME + "&cMusisz podac liczby!"));
@@ -46,19 +46,19 @@ public class GiveAkcesoriaCommand extends CommandAPI {
         switch (args[0]) {
             case "tarcza":
                 val3 = Integer.parseInt(args[3]);
-                player.getInventory().addItem(AkcesoriaPodsHelper.createTarcza(val1, val2, (int) val3, lvl, Utils.format(String.join(" ", Arrays.copyOfRange(args, 7, args.length)))));
+                player.getInventory().addItem(AkcesoriaPodsHelper.createTarcza(val1, val2, val3, lvl, Utils.format(String.join(" ", Arrays.copyOfRange(args, 7, args.length)))));
                 break;
             case "naszyjnik":
                 val1 = Integer.parseInt(args[1]);
-                player.getInventory().addItem(AkcesoriaPodsHelper.createNaszyjnik((int) val1, val2, val3, lvl, Utils.format(String.join(" ", Arrays.copyOfRange(args, 7, args.length)))));
+                player.getInventory().addItem(AkcesoriaPodsHelper.createNaszyjnik(val1, val2, val3, lvl, Utils.format(String.join(" ", Arrays.copyOfRange(args, 7, args.length)))));
                 break;
             case "kolczyki":
                 val3 = Integer.parseInt(args[3]);
-                player.getInventory().addItem(AkcesoriaPodsHelper.createKolczyki(val1, val2, (int) val3, lvl, Utils.format(String.join(" ", Arrays.copyOfRange(args, 7, args.length)))));
+                player.getInventory().addItem(AkcesoriaPodsHelper.createKolczyki(val1, val2, val3, lvl, Utils.format(String.join(" ", Arrays.copyOfRange(args, 7, args.length)))));
                 break;
             case "pierscien":
                 val3 = Integer.parseInt(args[3]);
-                player.getInventory().addItem(AkcesoriaPodsHelper.createPierscien(val1, val2, (int) val3, lvl, Utils.format(String.join(" ", Arrays.copyOfRange(args, 7, args.length)))));
+                player.getInventory().addItem(AkcesoriaPodsHelper.createPierscien(val1, val2, val3, lvl, Utils.format(String.join(" ", Arrays.copyOfRange(args, 7, args.length)))));
                 break;
             case "diadem":
                 player.getInventory().addItem(AkcesoriaPodsHelper.createDiadem(val1, val2, val3, lvl, Utils.format(String.join(" ", Arrays.copyOfRange(args, 7, args.length)))));
@@ -71,12 +71,12 @@ public class GiveAkcesoriaCommand extends CommandAPI {
                 break;
             case "medalion":
                 val2 = Integer.parseInt(args[2]);
-                player.getInventory().addItem(AkcesoriaDodatHelper.createMedalion(val1, (int) val2, lvl, Utils.format(String.join(" ", Arrays.copyOfRange(args, 7, args.length)))));
+                player.getInventory().addItem(AkcesoriaDodatHelper.createMedalion(val1, val2, lvl, Utils.format(String.join(" ", Arrays.copyOfRange(args, 7, args.length)))));
                 break;
             case "energia":
                 try {
-                    val3 = Double.parseDouble(args[3]);
-                    val4 = Double.parseDouble(args[4]);
+                    val3 = Integer.parseInt(args[3]);
+                    val4 = Integer.parseInt(args[4]);
                     val5 = Integer.parseInt(args[5]);
                 } catch (NumberFormatException e) {
                     player.sendMessage(Utils.format(Utils.SERVERNAME + "&cMusisz podac liczby!"));
